Divides gameControllers into general and actions Moves Lobby stuff from gameControllers into lobbyControllers Renames game and lobby routes and refactors fetch requests Refactored minor details Playing Draw2 + Draw4 skips the next user Added special chat logs for card actions
