+++ backend/controllers/gameControllers.js
@@ -1,384 +1,6 @@
 const db = require("../db/connection");
 const { StatusCodes } = require("http-status-codes");
 
-const logMessageToChat = (req, gameId, message) => {
-  req.app
-    .get("io")
-    .emit(`chat:message:${gameId}`, { from: "System", message: message });
-};
-
-const updateActiveSeat = async (userId, gameId) => {
-  const getTotalSeatsQuery = `SELECT COUNT(*) FROM game_users WHERE game_id = $1`;
-  const getCurrentSeatAndDirectionQuery = `SELECT game_users.seat, games.direction
-   FROM game_users
-   JOIN games ON game_users.game_id = games.id
-   WHERE game_users.game_id = $1 AND game_users.users_id = $2`;
-
-  const updateCurrentPlayerQuery = `UPDATE games
-   SET current_player_id =
-  (SELECT users_id FROM game_users WHERE seat = $1 AND game_id = $2)
-  WHERE id = $2
-  RETURNING current_player_id`;
-
-  const totalSeats = Number((await db.one(getTotalSeatsQuery, [gameId])).count);
-
-  const currentSeatAndDirection = await db.one(
-    getCurrentSeatAndDirectionQuery,
-    [gameId, userId],
-  );
-
-  const addend = currentSeatAndDirection.direction === "clockwise" ? 1 : -1;
-  let newSeat = currentSeatAndDirection.seat + addend;
-
-  if (newSeat < 1) {
-    newSeat += totalSeats;
-  } else {
-    newSeat = newSeat > totalSeats ? newSeat - totalSeats : newSeat;
-  }
-
-  return (await db.one(updateCurrentPlayerQuery, [newSeat, gameId]))
-    .current_player_id;
-};
-
-/**
- * Helper function that draws {drawNumber} cards
- *
- * @returns cards {
- *     color,
- *     symbol,
- *     id
- * }[]
- */
-const drawCards = async (currentPlayerId, gameId, drawNumber) => {
-  const getDeckCountQuery = `SELECT COUNT(*) FROM game_cards WHERE game_id = $1 AND user_id IS NULL`;
-
-  const deckCount = await db.oneOrNone(getDeckCountQuery, [gameId]);
-
-  if (!deckCount || Number(deckCount.count) < drawNumber) {
-    const getDiscardCountQuery = `SELECT COUNT(*) FROM game_cards WHERE game_id = $1 AND discarded = true`;
-    const discardCount = await db.oneOrNone(getDiscardCountQuery, [
-      gameId,
-      currentPlayerId,
-    ]);
-    if (!discardCount || discardCount.count < drawNumber) {
-      console.error("No more cards in discard pile or deck, ending game"); //TODO send message and skip player
-      return;
-    } else {
-      const restoreDeckQuery = `UPDATE game_cards SET user_id = NULL, discarded = false WHERE ( discarded = true AND game_id = $1 )`;
-      await db.none(restoreDeckQuery, [gameId]);
-    }
-  }
-
-  const selectIdsQuery = `SELECT card_id FROM game_cards WHERE user_id IS NULL AND game_id = $1 ORDER BY RANDOM() LIMIT $2 `;
-  const ids = await db.any(selectIdsQuery, [gameId, drawNumber]);
-  const cardIds = ids.map((id) => id.card_id);
-
-  const drawCardsQuery = `UPDATE game_cards SET user_id = $1 
-    WHERE game_id = $2 
-    AND card_id IN ($3:csv)`;
-
-  await db.any(drawCardsQuery, [currentPlayerId, gameId, cardIds]);
-
-  const getDrawnCardsQuery = `SELECT * FROM cards WHERE id IN ($1:csv)`;
-  const drawnCardsData = await db.any(getDrawnCardsQuery, [cardIds]);
-
-  const updatePlayerDeclaredUno =
-    "UPDATE game_users SET declared_uno = false WHERE game_id = $1 AND users_id = $2";
-  await db.none(updatePlayerDeclaredUno, [gameId, currentPlayerId]);
-  return drawnCardsData;
-};
-
-const reverseDirection = async (gameId) => {
-  const gameDirectionQuery = `UPDATE games
-   SET direction = (
-      CASE
-          WHEN (SELECT direction FROM games WHERE id = $1) = 'clockwise'::directions THEN 'counterclockwise'::directions
-          ELSE 'clockwise'::directions
-      END
-   )
-   WHERE id = $1`;
-
-  await db.none(gameDirectionQuery, [gameId]);
-};
-
-const isValidMove = async (color, symbol, gameId) => {
-  if (symbol === "wild" || symbol === "wild_draw_four") {
-    return true;
-  }
-
-  const activeCardAndColorQuery = `SELECT c.symbol, g.active_color 
-     FROM games g 
-     JOIN cards c ON g.current_card_id = c.id 
-     WHERE g.id = $1`;
-
-  const symbolAndColor = await db.one(activeCardAndColorQuery, [gameId]);
-  return (
-    symbolAndColor.active_color === color || symbolAndColor.symbol === symbol
-  );
-};
-
-const isOutOfTurn = async (gameId, userId) => {
-  const getActivePlayerQuery = `SELECT current_player_id FROM games WHERE id = $1`;
-  const activeId = await db.one(getActivePlayerQuery, [gameId]);
-  return activeId.current_player_id !== userId;
-};
-
-const drawCard = async (req, res) => {
-  const userId = req.session.user.id;
-  const gameId = req.params.id;
-  let activePlayerId, drawnCard, activePlayerHandSize;
-
-  if (await isOutOfTurn(gameId, userId)) {
-    console.error("Player is trying to move out of turn");
-    return res.status(400).send("It's not your turn, bucko");
-  }
-
-  try {
-    drawnCard = await drawCards(userId, gameId, 1);
-  } catch (error) {
-    console.error("Error in drawing card " + error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send("Error in drawing card " + error);
-  }
-
-  try {
-    activePlayerId = await updateActiveSeat(userId, gameId);
-    const getNextPlayerHandSize = `SELECT COUNT(*) FROM game_cards WHERE game_id = $1 AND user_id = $2 AND discarded = false`;
-    try {
-      activePlayerHandSize = await db.one(getNextPlayerHandSize, [
-        gameId,
-        activePlayerId,
-      ]);
-    } catch (err) {
-      console.error("error getting next player hand size ", err);
-      return res
-        .status(StatusCodes.INTERNAL_SERVER_ERROR)
-        .send(`Could not get next player hand size`);
-    }
-    req.app.get("io").to(gameId.toString()).emit("card-drawn", {
-      gameId: gameId,
-      clientId: userId,
-      activePlayerId: activePlayerId,
-      activePlayerHandSize: activePlayerHandSize.count,
-      drawnSymbol: drawnCard[0].symbol,
-      drawnColor: drawnCard[0].color,
-      drawnId: drawnCard[0].id,
-    });
-    logMessageToChat(req, gameId, `${req.session.user.username} drew a card`);
-    return res.status(StatusCodes.OK).send();
-  } catch (error) {
-    console.error("Error in updating active seat " + error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send("Error in updating active seat " + error);
-  }
-};
-
-const isWin = async (gameId, userId) => {
-  const getHandCount = `SELECT COUNT(card_id) FROM game_cards WHERE game_id = $1 AND user_id = $2 AND discarded = false`;
-  const handCount = await db.one(getHandCount, [gameId, userId]);
-  return handCount.count === "0";
-};
-
-const playCard = async (req, res) => {
-  let { color, symbol, isDeclared } = req.body;
-  const { id: gameId, cardId } = req.params;
-  const { id: userId } = req.session.user;
-
-  if (await isOutOfTurn(gameId, userId)) {
-    console.error("Player is trying to move out of turn");
-    return res.status(403).send("It's not your turn, bucko");
-  }
-
-  try {
-    const isValid = await isValidMove(color, symbol, gameId);
-    if (!isValid) {
-      return res.status(StatusCodes.FORBIDDEN).send("Player move not valid\n");
-    }
-  } catch (error) {
-    console.error("Could not determine if valid move " + error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send("Could not determine if valid move " + error);
-  }
-
-  const playCardQuery = `UPDATE games SET current_card_id = $1, active_color = $2 WHERE id = $3`;
-  const updatePlayerHandQuery = `UPDATE game_cards 
-  SET discarded = true
-  WHERE game_id = $1 AND card_id = $2 AND user_id = $3 RETURNING card_id`;
-  const updatePlayerDeclaredUno =
-    "UPDATE game_users SET declared_uno = true WHERE game_id = $1 AND users_id = $2";
-  try {
-    const card = await db.oneOrNone(updatePlayerHandQuery, [
-      gameId,
-      cardId,
-      userId,
-    ]);
-
-    //check if card played is in players hand
-    if (!card) {
-      return res.status(StatusCodes.FORBIDDEN).send("Don't cheat :/");
-    }
-
-    if (card.card_id !== Number(cardId)) {
-      return res
-        .status(StatusCodes.INTERNAL_SERVER_ERROR)
-        .send("Something went terribly wrong ;(");
-    }
-  } catch (error) {
-    console.error("Could not update player hand", error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not update player hand ${error}`);
-  }
-
-  try {
-    await db.none(playCardQuery, [cardId, color, gameId]);
-  } catch (error) {
-    console.error("Could not update active card", error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not update active card ${error}`);
-  }
-
-  if (symbol === "reverse") {
-    try {
-      await reverseDirection(gameId);
-    } catch (error) {
-      console.error("Error reversing direction", error);
-      return res
-        .status(StatusCodes.INTERNAL_SERVER_ERROR)
-        .send("Error reversing direction" + error);
-    }
-  }
-
-  if (isDeclared) {
-    try {
-      await db.none(updatePlayerDeclaredUno, [gameId, userId]);
-    } catch (error) {
-      console.error("Error updating player declared uno", error);
-      return res
-        .status(StatusCodes.INTERNAL_SERVER_ERROR)
-        .send("Error updating player declared uno" + error);
-    }
-  }
-
-  let activePlayerId;
-  try {
-    activePlayerId = await updateActiveSeat(userId, gameId);
-  } catch (error) {
-    console.error("Error in updating active seat " + error);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send("Error in updating active seat " + error);
-  }
-
-  let cards;
-  switch (symbol) {
-    case "draw_two":
-      try {
-        const getCurrentPlayerQuery = `SELECT current_player_id FROM games WHERE id = $1`;
-        const currentPlayerId = await db.one(getCurrentPlayerQuery, [gameId]);
-        cards = await drawCards(currentPlayerId.current_player_id, gameId, 2);
-        req.app.get("io").to(gameId.toString()).emit("cards-drawn", {
-          gameId: gameId,
-          cards,
-          currentPlayerId: currentPlayerId.current_player_id,
-        });
-      } catch (error) {
-        console.error("Error updating game_cards", error);
-        return res
-          .status(StatusCodes.INTERNAL_SERVER_ERROR)
-          .send("Error updating game_cards " + error);
-      }
-      break;
-    case "wild_draw_four":
-      try {
-        const getCurrentPlayerQuery = `SELECT current_player_id FROM games WHERE id = $1`;
-        const currentPlayerId = await db.one(getCurrentPlayerQuery, [gameId]);
-        cards = await drawCards(currentPlayerId.current_player_id, gameId, 4);
-
-        req.app.get("io").to(gameId.toString()).emit("cards-drawn", {
-          gameId: gameId,
-          cards,
-          currentPlayerId: currentPlayerId.current_player_id,
-        });
-      } catch (error) {
-        console.error("Error updating game_cards", error);
-        return res
-          .status(StatusCodes.INTERNAL_SERVER_ERROR)
-          .send("Error updating game_cards " + error);
-      }
-      break;
-    case "skip":
-      try {
-        const getCurrentPlayerQuery = `SELECT current_player_id FROM games WHERE id = $1`;
-        const nextPlayerId = (await db.one(getCurrentPlayerQuery, [gameId]))
-          .current_player_id;
-        activePlayerId = await updateActiveSeat(nextPlayerId, gameId);
-      } catch (error) {
-        console.error(
-          "Could not get total seats or current seat and direction",
-          error,
-        );
-        return res
-          .status(StatusCodes.INTERNAL_SERVER_ERROR)
-          .send(
-            "Could not get total seats or current seat and direction " + error,
-          );
-      }
-      break;
-    default:
-      break;
-  }
-  const getNextPlayerHandSize = `SELECT COUNT(*) FROM game_cards WHERE game_id = $1 AND user_id = $2 AND discarded = false`;
-  let nextPlayerHandSize;
-  try {
-    nextPlayerHandSize = await db.one(getNextPlayerHandSize, [
-      gameId,
-      activePlayerId,
-    ]);
-  } catch (err) {
-    console.error("error getting next player hand size ", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not get next player hand size`);
-  }
-
-  //update everyone in game with the new card played
-  req.app.get("io").to(gameId.toString()).emit("card-played", {
-    gameId,
-    color,
-    symbol,
-    clientId: userId,
-    cardId,
-    activePlayerId,
-    activePlayerHandSize: nextPlayerHandSize.count,
-  });
-  logMessageToChat(
-    req,
-    gameId,
-    `${req.session.user.username} played a ${color} ${symbol}`,
-  );
-  try {
-    if (await isWin(gameId, userId)) {
-      req.app
-        .get("io")
-        .to(gameId.toString())
-        .emit("is-win", { winnerName: req.session.user.username });
-      logMessageToChat(req, gameId, `${req.session.user.username} won!`);
-    }
-
-    return res.status(StatusCodes.OK).send("Success!");
-  } catch (err) {
-    console.error("error checking win condition ", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not check win condition`);
-  }
-};
-
 const getGameState = async (req, res) => {
   const { id: gameId } = req.params;
   const { id: userId } = req.session.user;
@@ -459,150 +81,6 @@ const getGameState = async (req, res) => {
   });
 };
 
-const joinGame = async (req, res) => {
-  const { id: gameId } = req.params;
-  const { password } = req.body;
-  const { id: userId } = req.session.user;
-
-  // Get the lobby if it exists { id, password, max players, player count }
-  const getLobbyQuery = `SELECT
-      g.id, g.password, g.max_players,
-      COUNT(gu.users_id) AS player_count
-    FROM games g
-    LEFT JOIN game_users gu ON g.id = gu.game_id
-    WHERE (g.id = $1)
-    GROUP BY g.id`;
-
-  // Get lobby id if user already in that lobby
-  const checkAlreadyJoinedQuery = `SELECT game_id FROM game_users
-    WHERE (game_id = $1)
-    AND (users_id = $2)`;
-
-  const addUserToLobby = `INSERT INTO game_users (game_id, users_id) VALUES ($1, $2)`;
-
-  // Check the lobby exists
-  let lobby;
-  try {
-    lobby = await db.oneOrNone(getLobbyQuery, [gameId]);
-
-    if (!lobby) {
-      return res.status(StatusCodes.NOT_FOUND).send(`Lobby does not exist`);
-    }
-
-    if (lobby.active) {
-      return res.redirect(`/game/${gameId}`);
-    }
-  } catch (err) {
-    console.error("error occurred getting lobby info ", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Server error while getting Lobby`);
-  }
-
-  if (lobby.password !== password && lobby.password !== null) {
-    return res
-      .status(StatusCodes.UNAUTHORIZED)
-      .send(`Incorrect password for Lobby`);
-  }
-
-  if (lobby.player_count >= lobby.max_players) {
-    return res.status(StatusCodes.CONFLICT).send(`Lobby is full`);
-  }
-
-  // Check user not already connected to lobby
-  try {
-    const isInLobby = await db.oneOrNone(checkAlreadyJoinedQuery, [
-      gameId,
-      userId,
-    ]);
-
-    // if user already in lobby, redirect to lobby page
-    if (isInLobby) {
-      req.session.errors = "User already in lobby";
-      return res.redirect(`/lobby/${gameId}`);
-    }
-  } catch (err) {
-    console.error("error occurred checking if user already in lobby", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Server error while checking if user already in lobby`);
-  }
-
-  // Add user to lobby
-  try {
-    await db.none(addUserToLobby, [gameId, userId]);
-    req.app
-      .get("io")
-      .to(gameId + "")
-      .emit("player-joined", {
-        username: req.session.user.username,
-        image: req.session.user.image,
-      });
-    logMessageToChat(
-      req,
-      gameId,
-      `${req.session.user.username} joined the game`,
-    );
-    return res.redirect(`/lobby/${gameId}`);
-  } catch (err) {
-    console.error("error occurred adding user to lobby", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Server error while adding user to lobby`);
-  }
-};
-
-const createGame = async (req, res) => {
-  const { name, max_players } = req.body;
-  let { password } = req.body;
-  const { id: userId } = req.session.user;
-
-  password = password === "" ? null : password;
-  const createLobbyQuery = `INSERT INTO games (name, password, max_players)
-      VALUES ($1, $2, $3)
-    RETURNING id`;
-
-  const joinLobbyQuery = `INSERT INTO game_users (users_id, game_id)
-    VALUES ($1, $2)`;
-
-  // Create a lobby
-  let lobby;
-  try {
-    lobby = await db.one(createLobbyQuery, [name, password, max_players]);
-  } catch (err) {
-    // If error was due to unique column value constraint for games.name
-    if (err.code === "23505") {
-      return res.status(StatusCodes.CONFLICT).send(`Lobby name taken`);
-    } else {
-      console.error("error creating lobby", err);
-      return res
-        .status(StatusCodes.INTERNAL_SERVER_ERROR)
-        .send(`User could not create lobby`);
-    }
-  }
-
-  // Add the user to the game
-  try {
-    const gameId = lobby.id;
-    await db.none(joinLobbyQuery, [userId, gameId]);
-
-    req.app.get("io").emit("lobby-created", {
-      id: gameId,
-      name,
-      max_players,
-      has_password: password != null,
-      player_count: 1,
-    });
-
-    return res.redirect(`/lobby/${gameId}`);
-  } catch (err) {
-    console.error("error adding user to lobby", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not add user to lobby`);
-  }
-};
-
 const startGame = async (req, res) => {
   const { id: gameId } = req.params;
   const { id: userId } = req.session.user;
@@ -644,70 +122,8 @@ const getUsersGames = async (req, res) => {
   }
 };
 
-const unoChallenge = async (req, res) => {
-  const userId = req.session.user.id;
-  const gameId = req.params.id;
-  const getCurrentPlayerQuery = `SELECT current_player_id FROM games WHERE id = $1`;
-  let currentPlayerId;
-  try {
-    (await db.one(getCurrentPlayerQuery, [gameId])).current_player_id;
-  } catch (err) {
-    console.error("error getting current player id", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not get current player id`);
-  }
-  const doesAnyoneHaveUnoQuery = `SELECT users_id 
-  FROM game_users 
-  WHERE game_id = $1
-  AND declared_uno = false
-  AND users_id IN
-  (SELECT user_id
-  FROM game_cards
-  WHERE game_id = $1
-    AND user_id != $2
-    AND discarded = false
-  GROUP BY user_id
-  HAVING COUNT(*) = 1);`;
-  let unoIds;
-  try {
-    unoIds = (await db.any(doesAnyoneHaveUnoQuery, [gameId, userId])).map(
-      (userId) => userId.users_id,
-    );
-  } catch (err) {
-    console.error("error getting uno ids", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send(`Could not get uno ids`);
-  }
-  let cards;
-  //make every player with uno draw 2 cards
-  try {
-    for (const id of unoIds) {
-      cards = await drawCards(id, gameId, 2);
-      req.app.get("io").to(gameId.toString()).emit("cards-drawn", {
-        gameId: gameId,
-        cards,
-        currentPlayerId: id,
-      });
-    }
-  } catch (err) {
-    console.error("Error updating game_cards", err);
-    return res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .send("Error updating game_cards " + err);
-  }
-
-  return res.status(StatusCodes.OK).send("Success!");
-};
-
 module.exports = {
-  playCard,
   getGameState,
-  joinGame,
-  drawCard,
-  createGame,
   getUsersGames,
   startGame,
-  unoChallenge,
 };
