+++ backend/routes/lobby.js
@@ -4,11 +4,17 @@ const {
   getLobbyUsers,
   getLobbies,
   getLobby,
+  joinLobby,
+  createLobby,
 } = require("../controllers/lobbyControllers");
 
 lobbyRouter.get("/:id/users", getLobbyUsers);
 
 lobbyRouter.get("/list", getLobbies);
+
+lobbyRouter.post("/:id/join", joinLobby);
+
+lobbyRouter.post("/create", createLobby);
 
 lobbyRouter.get("/:id", getLobby);
